health

+++ RenderQueue.py
@@ -35,7 +35,7 @@ class RenderQueue:
         screen.fill(background)
         for item in self.queue:
             if item.scale is not (1, 1):  # Scale if necessary
-                item.image = pygame.transform.scale(item.image, (item.image.get_width() * item.scale[0], item.image.get_height() * item.scale[1]))
+                item.image = pygame.transform.scale(item.image, ( int(item.image.get_width() * item.scale[0]), int(item.image.get_height() * item.scale[1])))
 
             # Draw the image
             screen.blit(item.image, item.loc)
